refactor: разделить каждый из контроллеров AdminController, PrivateController, PublicController на отдельные по функциям: Compilations, Events, Requests, Users
