+++ main-service/src/main/java/ru/practicum/mainservice/controllers/admincontrollers/CompilationAdminController.java
@@ -1,0 +1,51 @@
+package ru.practicum.mainservice.controllers.admincontrollers;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.*;
+import ru.practicum.mainservice.compilation.dto.CompilationDto;
+import ru.practicum.mainservice.compilation.dto.NewCompilationDto;
+import ru.practicum.mainservice.compilation.dto.UpdateCompilationRequest;
+import ru.practicum.mainservice.compilation.service.CompilationService;
+
+import javax.validation.Valid;
+import javax.validation.constraints.Positive;
+
+@RestController
+@RequestMapping(path = "/admin/compilations")
+@Slf4j
+@RequiredArgsConstructor
+@Validated
+public class CompilationAdminController {
+    private final CompilationService compilationService;
+
+    @PostMapping
+    public ResponseEntity<CompilationDto> createCompilation(@RequestBody @Valid NewCompilationDto newCompilationDto) {
+        log.info("POST-request was received at 'admin/compilations' . " +
+                "Create a COMPILATION: {}.", newCompilationDto);
+        return new ResponseEntity<>(compilationService.createCompilation(newCompilationDto), HttpStatus.CREATED);
+    }
+
+    @PatchMapping("/{compId}")
+    public ResponseEntity<CompilationDto> updateCompilation(
+            @PathVariable @Positive Long compId,
+            @RequestBody UpdateCompilationRequest updateCompilationDto) {
+        log.info("PATCH-request was received at 'admin/compilations/{}' . " +
+                "Update the COMPILATION with compId = {}. New DATA: {}.", compId, compId, updateCompilationDto);
+        return new ResponseEntity<>(compilationService.updateCompilation(compId, updateCompilationDto),
+                HttpStatus.OK);
+    }
+
+    @DeleteMapping("/{compId}")
+    public ResponseEntity<Void> deleteCompilation(@PathVariable @Positive Long compId) {
+        log.info("DELETE-request was received at 'admin/compilations/{}' . " +
+                "Delete a COMPILATION with compId = {}.", compId, compId);
+        compilationService.deleteCompilationById(compId);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    }
+
+
+}
