--- main-service/src/main/java/ru/practicum/mainservice/controllers/parameters/EventRequestSort.java
@@ -1,6 +1,0 @@
-package ru.practicum.mainservice.controllers.parameters;
-
-public enum EventRequestSort {
-    EVENT_DATE,
-    VIEWS
-}
