+++ main-service/src/main/java/ru/practicum/mainservice/controllers/publiccontrollers/parameters/EventRequestSort.java
@@ -1,0 +1,6 @@
+package ru.practicum.mainservice.controllers.publiccontrollers.parameters;
+
+public enum EventRequestSort {
+    EVENT_DATE,
+    VIEWS
+}
