+++ main-service/src/main/java/ru/practicum/mainservice/event/service/EventService.java
@@ -10,9 +10,9 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import ru.practicum.mainservice.category.model.Category;
 import ru.practicum.mainservice.category.repository.CategoryRepository;
-import ru.practicum.mainservice.controllers.parameters.EventAdminRequestParameters;
-import ru.practicum.mainservice.controllers.parameters.EventPublicRequestParameters;
-import ru.practicum.mainservice.controllers.parameters.EventRequestSort;
+import ru.practicum.mainservice.controllers.admincontrollers.parameters.EventAdminRequestParameters;
+import ru.practicum.mainservice.controllers.publiccontrollers.parameters.EventPublicRequestParameters;
+import ru.practicum.mainservice.controllers.publiccontrollers.parameters.EventRequestSort;
 import ru.practicum.mainservice.event.dto.*;
 import ru.practicum.mainservice.event.mapper.EventMapper;
 import ru.practicum.mainservice.event.model.*;
